api/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
black==23.12.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
datasets==2.18.0
dill==0.3.8
Django==5.0.1
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
filelock==3.13.3
frozenlist==1.4.1
fsspec==2024.2.0
huggingface-hub==0.22.2
idna==3.6
Markdown==3.5.2
mccabe==0.7.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
numpy==1.26.4
packaging==23.2
pandas==2.2.1
pathspec==0.12.1
platformdirs==4.1.0
psycopg2==2.9.9
pyarrow==15.0.2
pyarrow-hotfix==0.6
pycodestyle==2.11.1
pyflakes==3.2.0
python-dateutil==2.9.0.post0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
xxhash==3.4.1
yarl==1.9.4
